backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
qiskit==1.2.4
qiskit-aer==0.15.1
qiskit-machine-learning==0.7.2
google-generativeai==0.3.2
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0
